[P2PW-1217] - Working fee relayer topUp and send!

+++ solanaj/src/main/java/org/p2p/solanaj/model/types/TokenSupply.kt
@@ -20,7 +20,7 @@ class TokenSupplyValue(
     @SerializedName("decimals")
     val decimals: Int,
     @SerializedName("uiAmount")
-    val uiAmount: Long,
+    val uiAmount: Double,
     @SerializedName("uiAmountString")
     val uiAmountString: String,
 )
